refactor: common을 global 패키지의 하위패키지로 리팩토링
도메인에 종속되지 않는 일반 클래스에 대한 패키지 잘게 분리 위함
