+++ src/main/java/gift/global/common/annotation/PageParam.java
@@ -1,4 +1,4 @@
-package gift.common.annotation;
+package gift.global.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
