+++ src/main/java/gift/global/common/dto/SortInfo.java
@@ -1,4 +1,4 @@
-package gift.common.dto;
+package gift.global.common.dto;
 
 public record SortInfo(String field, boolean isAscending) {
